+++ tiger-proxy/src/test/java/de/gematik/test/tiger/proxy/TestTigerProxyTls.java
@@ -771,7 +771,7 @@ class TestTigerProxyTls extends AbstractTigerProxyTest {
   @SneakyThrows
   @Test
   void masterSecretFileDefined_shouldDumpSecretInCorrectForm() throws UnirestException {
-    final Path masterSecretsFile = Paths.get("target/masterSecrets.txt");
+    final Path masterSecretsFile = Paths.get("target/master-secrets.txt");
     FileUtils.deleteQuietly(masterSecretsFile.toFile());
 
     spawnTigerProxyWithDefaultRoutesAndWith(